# api/openapi.yaml
openapi: 3.0.3
info:
  title: Bloberry API
  version: 1.0.0
  description: |
    Storage-agnostic object service. One API, one SDK and one dashboard over
    S3, R2, OSS, GCS, Azure Blob or plain disk.

    Standard response envelope: `{data?, messages?: [{code, content}]}` — both
    `omitempty`, no `error` boolean. The HTTP status code is the success/failure
    signal; `messages` carries errors and confirmations. Two documented
    exceptions: `GET /s/{slug}` and public object URLs return HTML/redirects,
    not the envelope.

    Error codes are API — see the `ErrorCode` schema. Once shipped, a code
    never changes meaning.
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
  /auth/signup:
    post:
      operationId: signup
      summary: Sign up with an invitation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          $ref: '#/components/responses/AuthTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      operationId: login
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          $ref: '#/components/responses/AuthTokens'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /auth/refresh:
    post:
      operationId: refresh
      summary: Rotate a refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/AuthTokens'
        '401':
          $ref: '#/components/responses/ErrorResponse'
...
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      summary: Request a password reset link
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          $ref: '#/components/responses/MessageResponse'
  /auth/reset-password:
    post:
      operationId: resetPassword
      summary: Reset a password with a token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password reset
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/otp/request:
    post:
      operationId: requestOtp
      summary: Request an email OTP login code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          $ref: '#/components/responses/MessageResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/otp/verify:
    post:
      operationId: verifyOtp
      summary: Verify an OTP code and receive a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/AuthTokens'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /auth/google:
    post:
      operationId: googleLogin
      summary: Login with a Google id_token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_token]
              properties:
                id_token:
                  type: string
                platform:
                  $ref: '#/components/schemas/Platform'
      responses:
        '200':
          $ref: '#/components/responses/AuthTokens'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
...
  /auth/login/verify-totp:
    post:
      operationId: verifyTotpLogin
      summary: Complete a login blocked by totp_required
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pending, code]
              properties:
                pending:
                  type: string
                code:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/AuthTokens'
        '401':
          $ref: '#/components/responses/ErrorResponse'
...
  /auth/pair/verify:
    post:
      operationId: verifyPairToken
      summary: Exchange a pairing token for a mobile session (M22)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payload]
              properties:
                payload:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/AuthTokens'
        '400':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Human session JWT or a `blob_live_…` application access key. The
        middleware discriminates by prefix; everything downstream sees one
        principal.
...
  schemas:
    Platform:
      type: string
      enum: [web, mobile, desktop]
...
    SignupRequest:
      type: object
      required: [invite_token, email, password]
      properties:
        invite_token:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        display_name:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        platform:
          $ref: '#/components/schemas/Platform'
    Envelope:
      type: object
      properties:
        data: {}
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    Message:
      type: object
      required: [code]
      properties:
        code:
          type: string
        content:
          type: string
...
  responses:
    AuthTokens:
      description: Session tokens
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: object
                required: [access_token, refresh_token]
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  expires_in:
                    type: integer
                  totp_required:
                    type: boolean
                    description: set true when 2FA blocks session issue (M24)
    BadRequest:
      description: Validation or domain error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
    Conflict:
      description: Resource conflict (name_conflict, object_pending)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
    ErrorResponse:
      description: Error envelope with a documented code
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
    RateLimited:
      description: Rate limited (Retry-After header set)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
    MessageResponse:
      description: Confirmation message
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'